chore(API): Addressing review comments
